+++ src/parser/CMakeLists.txt
@@ -4,7 +4,10 @@ add_subdirectory(trees)
 add_subdirectory(io)
 add_subdirectory(tools)
 
-add_library(meta-parser sr_parser.cpp state.cpp transition.cpp)
+add_library(meta-parser sr_parser.cpp
+                        state.cpp
+                        state_analyzer.cpp
+                        transition.cpp)
 target_link_libraries(meta-parser meta-parser-trees
                                   meta-parser-io
                                   meta-util
